fixed face direction bug
