+++ DDF Project/Assets/Prefabs/PlayerGroup.prefab
@@ -27,11 +27,63 @@ Transform:
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
-  - {fileID: 6080464089295992697}
   - {fileID: 6080464089511710321}
+  - {fileID: 754153064629693449}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &5742337142235128070
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 754153064629693449}
+  - component: {fileID: 7346470698485036735}
+  m_Layer: 0
+  m_Name: GameObject
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &754153064629693449
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5742337142235128070}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: -0.252, y: 0.538, z: -1.648}
+  m_LocalScale: {x: 0.5, y: 0.5, z: 0.5}
+  m_Children:
+  - {fileID: 6080464089295992697}
+  m_Father: {fileID: 1893771849714761645}
+  m_RootOrder: 1
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &7346470698485036735
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5742337142235128070}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a346fd77f8bc750469cf05f1afb08bd8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  characterController: {fileID: 6080464089295992702}
+  MoveSpeed: 5
+  JumpSpeed: 20
+  RotationSpeed: 240
+  Gravity: 200
+  turnSmoothTime: 0.1
+  Music: {fileID: 0}
+  MusicJump: {fileID: 0}
 --- !u!1 &6080464088834122878
 GameObject:
   m_ObjectHideFlags: 0
@@ -88,7 +140,6 @@ GameObject:
   - component: {fileID: 6080464089295992697}
   - component: {fileID: 6080464089295992700}
   - component: {fileID: 6080464089295992699}
-  - component: {fileID: 6080464089295992701}
   - component: {fileID: 6080464089295992702}
   - component: {fileID: 6080464089295992703}
   m_Layer: 0
@@ -105,13 +156,13 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 6080464089295992696}
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -0.35, y: 0.53, z: -3.3}
-  m_LocalScale: {x: 0.5, y: 0.5, z: 0.5}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0.011408865, y: -0.016000032, z: -0.007999897}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 6080464090669968191}
   - {fileID: 6080464088834122879}
-  m_Father: {fileID: 1893771849714761645}
+  m_Father: {fileID: 754153064629693449}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!33 &6080464089295992700
@@ -161,22 +212,6 @@ MeshRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 0
---- !u!54 &6080464089295992701
-Rigidbody:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6080464089295992696}
-  serializedVersion: 2
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: 0.05
-  m_UseGravity: 1
-  m_IsKinematic: 0
-  m_Interpolate: 0
-  m_Constraints: 32
-  m_CollisionDetection: 0
 --- !u!143 &6080464089295992702
 CharacterController:
   m_ObjectHideFlags: 0
@@ -207,13 +242,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 615ddf8716458024eb0a228e4e2a698b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  characterController: {fileID: 0}
-  rb: {fileID: 0}
-  MoveSpeed: 4
-  JumpSpeed: 10
-  RotationSpeed: 240
-  Gravity: 20
-  turnSmoothTime: 0.1
+  Music: {fileID: 0}
+  MusicJump: {fileID: 0}
+  MusicRestore: {fileID: 0}
+  MusicDamage: {fileID: 0}
   Height: 0
   StartHealth: 4
   health: 0
@@ -251,7 +283,7 @@ Transform:
   m_LocalScale: {x: 0.5, y: 0.5, z: 0}
   m_Children: []
   m_Father: {fileID: 1893771849714761645}
-  m_RootOrder: 1
+  m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!33 &6080464089511710324
 MeshFilter:
