+++ DDF Project/Assets/Scripts/PlayerController.cs
@@ -7,23 +7,25 @@ using NaughtyAttributes;
 
 public class PlayerController : MonoBehaviour
 {
-    #region Movement Settings
-    [ReadOnly] [SerializeField] private CharacterController characterController;
-    [ReadOnly] [SerializeField] private Rigidbody rb;
-    public float MoveSpeed = 5.0f;
-    public float JumpSpeed = 20.0f;
-    public float RotationSpeed = 240.0f;
-    public float Gravity = 200.0f; // weight
-    private Vector3 moveDirection = Vector3.zero;
-    public float turnSmoothTime = 0.1f;
-    private float turnSmoothVelocity;
+    //#region Movement Settings
+    //[ReadOnly] [SerializeField] private CharacterController characterController;
+    //[ReadOnly] [SerializeField] private Rigidbody rb;
+
+    //public float MoveSpeed = 5.0f;
+    //public float JumpSpeed = 20.0f;
+    //public float RotationSpeed = 240.0f;
+    //public float Gravity = 200.0f; // weight
+    //private Vector3 moveDirection = Vector3.zero;
+    //private Vector3 moveDirection;
+    //public float turnSmoothTime = 0.1f;
+    //private float turnSmoothVelocity;
 
     public AudioSource Music;
     public AudioClip MusicJump;
     public AudioClip MusicRestore;
     public AudioClip MusicDamage;
 
-    #endregion
+    //#endregion
     [ReadOnly] public float Height;
     public int StartHealth = 4;
     [ReadOnly] [SerializeField] private int health;
@@ -31,40 +33,22 @@ public class PlayerController : MonoBehaviour
     [Required] public ShadowController ShadowController;
     private GameManager gameManager;
 
+    
+
     void Start()
     {
         gameManager = GameManager.Instance;
 
         health = StartHealth;
         gameManager.HealthBar.SetMaxHealth(health);
-        characterController = GetComponent<CharacterController>();
-        rb = GetComponent<Rigidbody>();
+        //characterController = GetComponent<CharacterController>();
+        //rb = GetComponent<Rigidbody>();
 
         Height = GetComponent<Renderer>().bounds.size.y;
+
+        
     }
 
-    void Update()
-    {
-        float horizontal = Input.GetAxisRaw("Horizontal");
-        float vertical = Input.GetAxisRaw("Vertical");
-        Vector3 move = new Vector3(horizontal, 0f, vertical).normalized;
-        moveDirection = move;
-
-        if (characterController.isGrounded && Input.GetButton("Jump"))
-        {
-            moveDirection.y = JumpSpeed;
-            Music.clip = MusicJump;
-            Music.Play();
-        }
-        moveDirection.y -= Gravity * Time.deltaTime;
-
-        // rotation
-        float targetAngle = Mathf.Atan2(moveDirection.x, moveDirection.z) * Mathf.Rad2Deg;
-        float angle = Mathf.SmoothDampAngle(transform.eulerAngles.y, targetAngle, ref turnSmoothVelocity, turnSmoothTime);
-        transform.rotation = Quaternion.Euler(0f, angle, 0f);
-
-        characterController.Move(moveDirection * MoveSpeed * Time.deltaTime);
-    }
 
 //    void OnCollisionEnter(Collision collision)
 //    {
